refactored discount method for tenpercent to double instead of int

+++ app/src/main/java/com/example/huascar/shoppingbasket/models/LoyaltyDiscount.java
@@ -8,7 +8,7 @@ import com.example.huascar.shoppingbasket.Interfaces.Discount;
 
 public class LoyaltyDiscount implements Discount {
 
-    public double Discount(int total) {
+    public double Discount(double total) {
         double result = total - (total*2)/100;
         return result;
     }
